src/api/requirements.txt
dnspython==2.2.1
fastapi==0.75.0
motor==2.5.1
python-dotenv==0.19.2
uvicorn==0.17.6